raspberry-pi/requirements.txt
aioice==0.9.0
aiortc==1.9.0
av==12.1.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
Flask==3.0.3
Flask-SocketIO==5.3.6
google-crc32c==1.5.0
h11==0.14.0
idna==3.7
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.10.0.82
pycparser==2.22
pyee==11.1.0
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.3
simple-websocket==1.0.0
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
